+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * POL_Wine_SelectPrefix: abort if no prefix name is provided
 * Sort install scripts lists case insensitively (iTunes)
 * Fix "dash-bang" line in shortcuts so they're actually executable
+* Deprecate the usage of $REPERTOIRE in PlayOnLinux own code
 
 # 4.2.6
 * Use $POL_TERM more consistently, allow POL_TERM global configuration